# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Marvel API",
    "description": "A service that returns data all about marvel comics"
  },
  "servers": [
    {
      "url": "http:\/\/localhost:8080",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/characters": {
      "get": {
        "summary": "Returns a list of characters",
        "description": "This api fetch data from Marvel Comics API(with cache implementation). ",
        "schema": {
          "$ref": "#/definitions/Characters"
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "An array of character IDs"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/characters/{id}": {
      "get": {
        "summary": "Returns a character details",
        "description": "This api fetch data from Marvel Comics API if id it not found in cache.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of character to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "schema": {
          "$ref": "#/definitions/Character"
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "An array of character IDs"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Character": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "Characters": {
      "type": "array"
    }
  }
}